KARTHIC-N-A (240524)/Assignment 1.1/requirements.txt
openai
numpy==2.2.6
pandas
python-dotenv
groq